1. 할당 최족화 2. 스케줄 기능 개발

+++ vo/JsonConfig.cs
@@ -7,12 +7,12 @@ namespace scte_104_inserter.vo
 	class JsonConfig
 	{
 		public String ipAddr { get; set; }
-		public Int32 port { get; set; }
+		public Int16 port { get; set; }
 		
-		public int eventId { get; set; }
-		public int uniquePid { get; set; }
-		public int preRollTime { get; set; }		
-		public int breakDuration { get; set; }
+		public Int32 eventId { get; set; }
+		public Int16 uniquePid { get; set; }
+		public Int16 preRollTime { get; set; }		
+		public Int16 breakDuration { get; set; }
 
 		public String configFileName = "config.json";
